feat: Add pagination to user list page
This commit adds pagination functionality to the user list page,
allowing users to view a large number of users in smaller, more
manageable groups. The pagination controls are located at the bottom
of the page and allow the user to navigate through the different pages.

The following changes were made:
- Added pagination functionality to user list page
- Updated styling for pagination controls